Refactor statistics and types
Make Type::CloneType abstract. Change unsigned to size_t in some
places in namespace model. Add keyword override for overrided methods.
Refactor ToKeyValueStringPairs to ToKeyValueMap.Change CsvStats::GetMin
signature, remove GetMinMax_. Rename CsvStats::STDAndCentralMoment_ to
CsvStats::CalculateCentralMoment. Refactor CsvStats::ShowSample and
it's description. Add GetValue to TypedColumnData. Add IsOrdered static
method to Type. Add CreateColumnData static methods to FDAlgorithm.

+++ src/model/relation_data.h
@@ -26,21 +26,21 @@ public:
                                   std::vector<ColumnType> column_data) noexcept
         : schema_(std::move(schema)), column_data_(std::move(column_data)) {}
 
-    virtual unsigned int GetNumRows() const = 0;
+    virtual size_t GetNumRows() const = 0;
 
     virtual std::vector<ColumnType>& GetColumnData() { return column_data_; }
     virtual std::vector<ColumnType> const& GetColumnData() const { return column_data_; }
-    virtual ColumnType& GetColumnData(unsigned int column_index) {
+    virtual ColumnType& GetColumnData(size_t column_index) {
         return column_data_[column_index];
     }
-    virtual ColumnType const& GetColumnData(unsigned int column_index) const {
+    virtual ColumnType const& GetColumnData(size_t column_index) const {
         return column_data_[column_index];
     }
 
     virtual ~AbstractRelationData() = default;
 
     double GetMaximumEntropy() const { return std::log(GetNumRows()); }
-    unsigned int GetNumColumns() const { return schema_->GetNumColumns(); }
+    size_t GetNumColumns() const { return schema_->GetNumColumns(); }
     double GetMaximumNip() const { return GetNumRows() * (GetNumRows() - 1) / 2.0; }
     unsigned long long GetNumTuplePairs() const {
         return (unsigned long long)GetNumRows() * (GetNumRows() - 1) / 2;
